Share the chat loop-prevention test fixtures
The reply constants, seeded history and recording backend were copy-pasted
across five test modules, and had already drifted: the retry test's
LOOPED_REPLY dropped a paragraph, so it exercised the similarity thresholds
against ~130 fewer characters than every other copy. The recording backends
had also diverged in parameter order from the real
RemoteModelLike.generate_chat_response signature they exist to police --
that signature match is the only thing standing between a renamed kwarg and
a silently-passing suite.

tests/chat_fixtures.py now holds LOOPED_REPLY, FRESH_REPLY,
SECOND_OPINION_REPLY, the mandated Medicaid block, and RecordingChatModel;
the five modules import them. Raised by CodeRabbit on the original review
and left unaddressed until now.
